Rename config classes
