+++ src/test/kotlin/de/bund/digitalservice/useid/journey/JourneyTestConfig.kt
@@ -1,8 +1,8 @@
-package de.bund.digitalservice.useid.config
+package de.bund.digitalservice.useid.journey
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.boot.test.context.TestConfiguration
 
 @TestConfiguration
-@EnableConfigurationProperties(TestApplicationProperties::class)
-class TestConfig
+@EnableConfigurationProperties(JourneyTestApplicationProperties::class)
+class JourneyTestConfig
